+++ src/App.jsx
@@ -4,10 +4,10 @@ import FormAuthorization from "./Components/FormAuthorization/FormAuthorization"
 import { useContext, useEffect } from "react";
 import "./App.css";
 import ModalPortal from "./Components/ModalPortal/ModalPortal";
-import { UserContext } from "./Components/Context/Context";
+import { UserContext } from "./Context/Context";
 import { useNavigate, Outlet, useLocation } from "react-router-dom";
 import { Button } from "antd";
-import { apiUser } from "./Components/Api/User";
+import { apiUser } from "./Api/User";
 
 function App() {
   const {
@@ -31,7 +31,6 @@ function App() {
       setUserToken(token);
       apiUser.getInfoUser().then((user) => {
         setUser(user);
-        console.log(user);
       });
     } else if (!isAuth && location.pathname !== "/signup") {
       setModalOpen(true);
